backend/requirements.txt
fastapi==0.121.2
uvicorn==0.38.0
pydantic==2.12.4
python-multipart==0.0.20

# 数据处理及工具
pandas==2.2.3
pyparsing==3.2.5
jieba==0.42.1
python-dotenv==1.2.1
requests==2.32.5

# 向量数据库
chromadb==1.3.5

# AI 框架及 SDK
llama-index==0.14.8
llama-index-retrievers-bm25==0.6.5
zai-sdk==0.0.4.2